Fix resize parser

+++ noise_argparser.py
@@ -47,8 +47,8 @@ def parse_dropout(dropout_command):
 def parse_resize(resize_command):
     matches = re.match(r'resize\((\d+\.*\d*,\d+\.*\d*)\)', resize_command)
     ratios = matches.groups()[0].split(',')
-    min_ratio = int(ratios[0])
-    max_ratio = int(ratios[1])
+    min_ratio = float(ratios[0])
+    max_ratio = float(ratios[1])
     return Resize((min_ratio, max_ratio))
